device-abstraction restructure, pr-extra changes
