+++ src/arcana-incubator/pr-util/framegraph/types.hh
@@ -18,8 +18,8 @@ inline constexpr physical_res_idx gc_invalid_physical_res = uint32_t(-1);
 
 struct res_handle // long name: virtual resource handle
 {
-    virtual_res_idx resource;
-    int version_before;
+    virtual_res_idx resource = gc_invalid_virtual_res;
+    int version_before = -1;
 };
 
 struct access_mode
